Move CustomText to correct folder
